requirements.txt

jsonschema>=4.22.0
aiomqtt>=2.0.0
protobuf>=4.25.0
fastapi>=0.110.0
uvicorn>=0.23.0
redis>=5.0.0
PyJWT>=2.8.0
